fixed timecode for cDNG header, will iwork if whole clip exported, if cut in is > 0 the first DNG will have cutin - 1 frames's time code, will not work with cut audio

+++ src/dng/dng.h
@@ -57,7 +57,7 @@ int dng_decompress_image(uint16_t * output_buffer, uint16_t * input_buffer, size
 
 /* routines to initialize, save and free DNG exporting struct */
 dngObject_t * initDngObject(mlvObject_t * mlv_data, int raw_state, double fps, int32_t par[4]);
-int saveDngFrame(mlvObject_t * mlv_data, dngObject_t * dng_data, uint64_t frame_index, char * dng_filename);
+int saveDngFrame(mlvObject_t * mlv_data, dngObject_t * dng_data, uint32_t frame_index, char * dng_filename);
 void freeDngObject(dngObject_t * dng_data);
 
 #endif
